tests: group wrappers & make them optional
