+++ tests/unittests/multimodal/test_clip_iqa.py
@@ -26,9 +26,9 @@ from torchmetrics.multimodal.clip_iqa import CLIPImageQualityAssessment
 from torchmetrics.utilities.imports import _PIQ_GREATER_EQUAL_0_8, _TRANSFORMERS_GREATER_EQUAL_4_10
 from torchvision.transforms import PILToTensor
 
+from unittests.helpers import skip_on_connection_issues
 from unittests.helpers.testers import MetricTester
 from unittests.image import _SAMPLE_IMAGE
-from unittests.text.helpers import skip_on_connection_issues
 
 
 @pytest.mark.parametrize(
