+++ tests/unittests/text/test_bertscore.py
@@ -22,7 +22,8 @@ from torchmetrics.text.bert import BERTScore
 from torchmetrics.utilities.imports import _BERTSCORE_AVAILABLE, _TRANSFORMERS_GREATER_EQUAL_4_4
 from typing_extensions import Literal
 
-from unittests.text.helpers import TextTester, skip_on_connection_issues
+from unittests.helpers import skip_on_connection_issues
+from unittests.text.helpers import TextTester
 from unittests.text.inputs import _inputs_single_reference
 
 if _BERTSCORE_AVAILABLE:
