+++ tests/unittests/text/test_rouge.py
@@ -24,7 +24,8 @@ from torchmetrics.text.rouge import ROUGEScore
 from torchmetrics.utilities.imports import _NLTK_AVAILABLE, _ROUGE_SCORE_AVAILABLE
 from typing_extensions import Literal
 
-from unittests.text.helpers import TextTester, skip_on_connection_issues
+from unittests.helpers import skip_on_connection_issues
+from unittests.text.helpers import TextTester
 from unittests.text.inputs import _Input, _inputs_multiple_references, _inputs_single_sentence_single_reference
 
 if _ROUGE_SCORE_AVAILABLE:
